DrawLib change.
The analog clock before only represented the time as a fraction of 60 minutes.
The time is now represented as the full clock and the time moves faster, it also runs counter clockwise instead of clockwise.
It was done this way because seconds could not be properly represented using the original,
this way you get a btter impression of how much time is left.