Child name is now properly shown in header.
The problem was fixed by fixing the context to 'this',
by assigning the childProfile a null value if timer is not started from the launcher(e.g no extras are sent over)
and by checking the value of childProfile and not using the childprofile if it is null.

Consider not allowing opening the app outside of the Launcher at all.